# [Perf] Adjust KV Cache for torch.compile friendly