--- a/dia/state.py
+++ b/dia/state.py
@@ -58,6 +58,7 @@ class EncoderInferenceState:
             torch.arange(config.data.text_length, dtype=torch.float32, device=device).unsqueeze(0).expand(2, -1)
         )
         padding_mask = (cond_src != config.data.text_pad_value).to(device).expand(2, -1)
+
         attn_mask = create_attn_mask(padding_mask, padding_mask, device, is_causal=False)
 
         return cls(
@@ -69,7 +70,7 @@ class EncoderInferenceState:
         )
 
 
-class KVCache:
+class KVCache(torch.nn.Module):
     def __init__(
         self,
         num_heads: int,
@@ -80,9 +81,12 @@ class KVCache:
         k: torch.Tensor | None = None,
         v: torch.Tensor | None = None,
     ):
-        self.k = torch.zeros((2, num_heads, max_len, head_dim), dtype=dtype, device=device) if k is None else k
-        self.v = torch.zeros((2, num_heads, max_len, head_dim), dtype=dtype, device=device) if v is None else v
-        self.current_idx = 0
+        k = torch.zeros((2, num_heads, max_len, head_dim), dtype=dtype, device=device) if k is None else k
+        v = torch.zeros((2, num_heads, max_len, head_dim), dtype=dtype, device=device) if v is None else v
+
+        super().__init__()
+        self.register_buffer("k", k)
+        self.register_buffer("v", v)
 
     @classmethod
     def from_kv(cls, k: torch.Tensor, v: torch.Tensor) -> "KVCache":
@@ -96,11 +100,13 @@ class KVCache:
             v=v,
         )
 
-    def update(self, k: torch.Tensor, v: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
-        self.k[:, :, self.current_idx : self.current_idx + 1, :] = k
-        self.v[:, :, self.current_idx : self.current_idx + 1, :] = v
-        self.current_idx += 1
-        return self.k[:, :, : self.current_idx, :], self.v[:, :, : self.current_idx, :]
+    def update(self, k: torch.Tensor, v: torch.Tensor, current_idx: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
+        k_out, v_out = self.k, self.v
+        k_out[:, :, current_idx, :] = k
+        v_out[:, :, current_idx, :] = v
+        # self.current_idx += 1
+        # return self.k[:, :, : self.current_idx, :], self.v[:, :, : self.current_idx, :]
+        return self.k, self.v
 
     def prefill(self, k: torch.Tensor, v: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
         prefill_len = k.shape[2]
@@ -121,6 +127,7 @@ class DecoderInferenceState:
     dec_cross_attn_mask: torch.Tensor
     self_attn_cache: list[KVCache]
     cross_attn_cache: list[KVCache]
+    casual_attn_mask: torch.Tensor
 
     @classmethod
     def new(
@@ -137,7 +144,9 @@ class DecoderInferenceState:
 
         dec_positions = torch.full((2, 1), fill_value=0, dtype=torch.int32, device=device)
         tgt_padding_mask = torch.ones((2, 1), dtype=torch.bool, device=device)
+        
         dec_cross_attn_mask = create_attn_mask(tgt_padding_mask, enc_state.padding_mask, device, is_causal=False)
+        causal_mask = torch.tril(torch.ones(max_audio_len, max_audio_len, dtype=torch.bool, device=device))
 
         self_attn_cache = [
             KVCache(
@@ -159,6 +168,7 @@ class DecoderInferenceState:
             dec_cross_attn_mask=dec_cross_attn_mask,
             self_attn_cache=self_attn_cache,
             cross_attn_cache=dec_cross_attn_cache,
+            casual_attn_mask=causal_mask,
         )
 
     def prepare_step(self, step_from: int, step_to: int | None = None) -> None: