@JakubCzarlinski 2025-05-02
@tsdocode Amazing work here!

Tested this after the merge. Another 30 tokens/s compared to #162 in my env.

```
generate step 344: speed=234.115 tokens/s, realtime factor=2.722x
generate step 430: speed=233.398 tokens/s, realtime factor=2.714x
generate step 516: speed=234.825 tokens/s, realtime factor=2.731x
generate step 602: speed=233.978 tokens/s, realtime factor=2.721x
generate step 688: speed=233.303 tokens/s, realtime factor=2.713x
generate step 774: speed=233.219 tokens/s, realtime factor=2.712x
```